requirements.txt
numpy>=1.21.0
pandas>=1.3.0
gymnasium>=0.28.1
stable-baselines3[extra]>=2.0.0
torch>=2.0.0
matplotlib>=3.4.0
pytest>=7.0.0
scipy>=1.7.0
entsoe-py>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.18.0
PyYAML>=6.0

